solved the DOM issue. The score is still not updating
