+++ functions/palabraExiste.js
@@ -1,7 +1,4 @@
-export default function palabraExiste(palabra, lista, diccionario) {
+export default function palabraExiste(palabra, diccionario) {
     //esto debe revisar si la palabra está en el diccionario Y si no se escribió ya anteriormente.
-    let boo = diccionario.includes(palabra.join("").toLowerCase())
-    lista.includes(palabra.join("").toLowerCase()) && 
-    !boo ? false : true; 
-    //aquí invertí los valores falso y verdadero, para que sea más intuitivo
+    return diccionario.includes(palabra.join("").toLowerCase()) ? true: false;
   }
